feat(approval): visualize SQL risk analysis progress
- persist SQL recognition, behavior analysis, and security rule activities with independent lifecycle and progress
- expose analysis status, results, processed counts, and elapsed time in ticket details
- mark the current stage as failed and close subsequent stages when ticket processing fails
- size approval workers from available CPUs and remove the legacy RDP async queue dependency
- remove obsolete approval analysis columns and simplify ticket persistence and DTO mappings

+++ backend/clouddm-platform/cgdm-console/src/main/java/com/clougence/clouddm/console/web/model/vo/ticket/DmQueryTicketVO.java
@@ -28,23 +28,23 @@ import lombok.Setter;
 @Setter
 public class DmQueryTicketVO {
 
-    private Long                   id;
-    private String                 bizId;
-    private String                 gmtCreate;
-    private String                 gmtModified;
-    private Long                   dataSourceId;
-    private SqlContentType         contentType;
-    private Long                   attachmentId;
-    private String                 attachmentFileName;
-    private Long                   attachmentFileSize;
-    private List<ApprovalBehavior> behaviors;
-    private Integer                totalCount;
-    private String                 description;
-    private String                 statusMessage;
-    private Long                   expectedAffectedRows;
-    private Boolean                immediately;
-    private String                 rollBackSql;
-    private String                 ticketMessage;
-    private boolean                autoExec;
+    private Long                        id;
+    private String                      bizId;
+    private String                      gmtCreate;
+    private String                      gmtModified;
+    private Long                        dataSourceId;
+    private SqlContentType              contentType;
+    private Long                        attachmentId;
+    private String                      attachmentFileName;
+    private Long                        attachmentFileSize;
+    private List<ApprovalBehavior>      behaviors;
+    private Long                        totalCount;
+    private String                      description;
+    private String                      statusMessage;
+    private Long                        expectedAffectedRows;
+    private Boolean                     immediately;
+    private String                      rollBackSql;
+    private String                      ticketMessage;
+    private boolean                     autoExec;
     private List<TicketRuleCheckResult> checkedList;
 }
